[Docs Site] Fix index.md encoding on /plans/
